requirements.txt
paho-mqtt==1.5.0
Flask==1.1.1
